[Gradle, JS] Move JsCompilerType to Gradle Plugin API
