+++ libraries/tools/kotlin-gradle-plugin-api/src/main/kotlin/org/jetbrains/kotlin/gradle/plugin/JsCompilerType.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.gradle.targets.js
+package org.jetbrains.kotlin.gradle.plugin
 
 import org.gradle.api.Named
 import java.io.Serializable
